+++ test/ring_logger/client_test.exs
@@ -4,7 +4,15 @@ defmodule RingLogger.Client.Test do
 
   setup do
     {:ok, client} = Client.start_link()
+    # Elixir 1.4 changed the default pattern (removed $levelpad) so hardcode a default
+    # pattern here
+    Client.configure(client, format: "\n$time $metadata[$level] $message\n")
     {:ok, %{client: client}}
+  end
+
+  setup(context) do
+    RingLogger.ApplicationEnvHelpers.with_application_env(context, &on_exit/1)
+    :ok
   end
 
   describe "configure a client at runtime" do
@@ -80,7 +88,7 @@ defmodule RingLogger.Client.Test do
   test "can retrieve configuration", %{client: client} do
     config = [
       colors: %{debug: :cyan, enabled: true, error: :red, info: :normal, warn: :yellow},
-      format: ["\n", :time, " ", :metadata, "[", :level, "] ", :levelpad, :message, "\n"],
+      format: ["\n", :time, " ", :metadata, "[", :level, "] ", :message, "\n"],
       io: :stdio,
       level: :debug,
       metadata: [],
@@ -88,5 +96,29 @@ defmodule RingLogger.Client.Test do
     ]
 
     assert Client.config(client) == config
+  end
+
+  @tag application_envs: [ring_logger: [colors: %{debug: :green, error: :blue}]]
+  test "warns on deprecated config" do
+    io_warning =
+      ExUnit.CaptureIO.capture_io(:stderr, fn ->
+        Client.start_link()
+      end)
+
+    assert io_warning =~ ~r/:ring_logger.*is deprecated/
+  end
+
+  @tag application_envs: [
+         ring_logger: [colors: %{debug: :green, error: :blue}],
+         logger: [{RingLogger, [colors: %{debug: :cyan}]}]
+       ]
+  test "non-deprecated config override deprecated config" do
+    # Note: `error: :blue` in configuration will be overwritten because maps are not deep merged
+    ExUnit.CaptureIO.capture_io(:stderr, fn ->
+      {:ok, client} = Client.start_link()
+
+      assert %{debug: :cyan} = :sys.get_state(client).colors
+      assert :sys.get_state(client).level == :debug
+    end)
   end
 end